Start Record Collection Challenge
In this challenge, I am writing a function that allows the user to
manipulate data in a nested JavaScript Object of a record collection.

I have started the updateRecords function, the first argument to check
is if the prop passed into the function is not "tracks" and the value
passed in is not null then we will update the value for the record's
property by using the id passed into the function to decide which record
we will be updating.