Add additional if statement
I added another statement which checks the parameters for a prop that
equals tracks but the id for the album passed in does NOT have a tracks
property, the code block executed will create a tracks array and then
push the value argument into the array.  This has been tested and works.