Add final two conditional statements
I had to modify the if prop is "tracks" and value is not equal to "" in
order to execute an argument where tracks array was there but the
statement was adding the empty string to the array.  It now passes this
conditional statement and goes to if value = empty("") the code will
delete the property that the user passes in.