Create file profileLookup.js
This challenge involves writing a function lookupProfiles that takes in
two parameters name and props and iterates through the array objects
representing different people in our contacts list.

I have written the pseudocode for the if statements to make the function
work properly.  I need to figure out which loop to implement when
working with an array of objects.