Function checking if name and prop exists
Using a basic for loop to iterate over the contacts object, I then
nested a conditional statement which checks to see if the parameters
exist in the object.  This works better than my previous conditional
statement at the beginning of the function which only checked if a
firstName property exists and if the prop hasOwnProperty which always returns
true and is misleading.