Function sucessfully returns requested data
The lookUpProfile function takes in two parameters (name, prop).  Then
it checks the object to make sure the name exists in contacts.  If it
does it checks the prop argument and prints those to the console.  If
the name is not found, it logs "No such contact" to the console.

TODO: Find a way to not print No such contact when a contact DOES
exist.  Right now, the function works, but always prints 'No such
contact' even when the desired contact is present.