Use multiple conditional statements
To evaluate the user's strokes against the par of the hole, the numbers
are passed into a function called golfScore.  Multiple conditional
statments will be used to determine what index in the names array to
print to the console.

The names array holds multiple strings each one represent a score a
player can attain in a hole of golf.